backend/requirements.txt
django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
python-dotenv==1.1.1
firebase-admin==7.3.0
openai==2.31.0
google-generativeai==0.8.6
anthropic==0.94.0
requests==2.32.3
Pillow==11.2.1
gunicorn==25.3.0
